restructuring polls
