+++ src/Nav.js
@@ -7,9 +7,9 @@ import Family from './Family';
 import Login from './Login';
 import SignUp from './SignUp';
 import Mood from './Mood';
-import AllPolls from './AllPolls';
-import SinglePoll from './SinglePoll';
-import CreatePoll from './CreatePoll';
+import AllPolls from './Polls/AllPolls';
+import SinglePoll from './/Polls/SinglePoll';
+import CreatePoll from './Polls/CreatePoll';
 import ForgotPassword from './ForgotPassword';
 import Events from './Events';
 import SingleEvent from './SingleEvent';
